Add files via upload
A method that accepts a value and creates a new node, assigns it to the list head, and returns a pointer to the new head node.
A method to remove the head node and return the new list head node. If the list is empty, returns null.
A method to return the value (not the node) at the head of the list. If the list is empty, returns null.
A display() method that uses a while loop and a runner to return a string containing all list values.